+++ Core/MIPS/InterpreterVFPU.cpp
@@ -31,7 +31,7 @@
 #include "Core/System.h"
 
 #include "Core/MIPS/MIPS.h"
-#include "Core/MIPS/MIPSInt.h"
+#include "Core/MIPS/Interpreter.h"
 #include "Core/MIPS/MIPSTables.h"
 #include "Core/MIPS/MIPSVFPUUtils.h"
