+++ Core/MIPS/MIPSTables.cpp
@@ -22,8 +22,8 @@
 #include "Core/MIPS/MIPS.h"
 #include "Core/MIPS/MIPSDis.h"
 #include "Core/MIPS/MIPSDisVFPU.h"
-#include "Core/MIPS/MIPSInt.h"
-#include "Core/MIPS/MIPSIntVFPU.h"
+#include "Core/MIPS/Interpreter.h"
+#include "Core/MIPS/InterpreterVFPU.h"
 #include "Core/MIPS/MIPSCodeUtils.h"
 #include "Core/MIPS/MIPSTables.h"
 #include "Core/CoreTiming.h"
