+++ android/jni/Android.mk
@@ -514,8 +514,8 @@ EXEC_AND_LIB_FILES := \
   $(SRC)/Core/MIPS/MIPSDis.cpp \
   $(SRC)/Core/MIPS/MIPSDisVFPU.cpp \
   $(SRC)/Core/MIPS/MIPSAsm.cpp \
-  $(SRC)/Core/MIPS/MIPSInt.cpp.arm \
-  $(SRC)/Core/MIPS/MIPSIntVFPU.cpp.arm \
+  $(SRC)/Core/MIPS/Interpreter.cpp.arm \
+  $(SRC)/Core/MIPS/InterpreterVFPU.cpp.arm \
   $(SRC)/Core/MIPS/MIPSStackWalk.cpp \
   $(SRC)/Core/MIPS/MIPSTables.cpp \
   $(SRC)/Core/MIPS/MIPSVFPUUtils.cpp.arm \
